pyproject.toml
[tool.poetry]
name = "amam"
version = "1.1"
description = ""
authors = ["Hoai Tu <khaihoai1107@gmail.com>"]
readme = "README.md"
packages = [{include = "amam"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.89.1"}
sqlalchemy = "^1.4.46"
psycopg2 = "^2.9.5"
black = "^22.12.0"
alembic = "^1.9.2"
bcrypt = "^4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:start"
